BLOC-29 - Created generic.js and implemented common functions in this js
