+++ scripts/profile.js
@@ -1,16 +1,13 @@
+import {
+    redirectToLogin,
+    hideFullScreenLoader,
+    showFullScreenLoader,
+    loadNavbarAndSidebar,
+    truncateWalletAddress,
+    handleLogout,
+    handleCopyWalletAddress,
+} from './generic.js';
 
-// let currentPage = 1;
-// let totalPages = 1;
-// let query = '';
-// let filter = '';
-// const prevButton = document.getElementById("prev-button");
-// const nextButton = document.getElementById("next-button");
-
-// // Redirect to login if no token or token is invalid
-function redirectToLogin() {
-    chrome.storage.sync.remove(['authToken', 'connectedSites', 'email']);
-    window.location.href = 'login.html';
-}
 function formatAmount(amount) {
     // Convert the amount to a number
     amount = parseFloat(amount);
@@ -29,15 +26,7 @@ function formatAmount(amount) {
         return amount.toFixed(2); // If it's less than 1,000, show the number with two decimals
     }
 }
-// Function to show the full screen loader
-function showFullScreenLoader() {
-    document.getElementById('full-screen-loader').style.display = 'flex';
-}
 
-// Function to hide the full screen loader
-function hideFullScreenLoader() {
-    document.getElementById('full-screen-loader').style.display = 'none';
-}
 // // Fetch updated user profile from the API
 async function fetchUpdatedUserProfile() {
     showFullScreenLoader();
@@ -76,58 +65,6 @@ async function fetchUpdatedUserProfile() {
     }
 }
 
-function truncateWalletAddress(walletAddress, startChars = 6, endChars = 6, separator = '.......') {
-    if (!walletAddress || walletAddress.length <= startChars + endChars) {
-        return walletAddress; // Return the full address if it's too short to truncate
-    }
-    return `${walletAddress.substring(0, startChars)}${separator}${walletAddress.substring(walletAddress.length - endChars)}`;
-}
-
-// // Lock wallet and redirect to login
-async function lockWallet() {
-    showFullScreenLoader();
-    const { authToken } = await chrome.storage.sync.get('authToken');
-    const { email } = await chrome.storage.sync.get('email');
-    if (!authToken) {
-        console.error('No authToken found. Cannot log out.');
-        hideFullScreenLoader();
-        return;
-    }
-
-    try {
-        const response = await fetch(`https://dev-wallet-api.dubaicustoms.network/api/ext-logout?email=${encodeURIComponent(email)}`, {
-            method: 'GET',
-            headers: { 'Authorization': `Bearer ${authToken}` }
-        });
-
-        if (response.ok) {
-            const data = await response.json();
-            hideFullScreenLoader();
-            if (data.message === "Successfully Logged Out") {
-                chrome.storage.sync.remove(['authToken', 'connectedSites', 'email'], () => {
-                    chrome.runtime.sendMessage({ action: 'lock_wallet' }, (response) => {
-                        if (response.success) {
-                            window.location.href = 'login.html';
-                            hideFullScreenLoader();
-                        } else {
-                            console.error('Failed to close full-screen tab.');
-                            hideFullScreenLoader();
-                        }
-                    });
-                });
-                chrome.runtime.sendMessage({ action: 'logout' });
-            } else {
-                alert('Logout failed. Please try again.');
-            }
-        } else {
-            alert('Logout failed. Please try again.');
-        }
-    } catch (error) {
-        hideFullScreenLoader();
-        console.error('Error during logout:', error);
-        alert('An error occurred during logout. Please try again.' + response.status);
-    }
-}
 //   // Event listener for DOM content loading
   document.addEventListener('DOMContentLoaded', async () => {
     showFullScreenLoader();
@@ -144,26 +81,7 @@ async function lockWallet() {
     const balanceElement = document.getElementById('balance');
     const emailElement = document.getElementById('email');
 
-    const copyButton = document.getElementById('copy-button');
-    const copyMessage = document.getElementById('copy-message');
-    if(copyButton){
-        copyButton.addEventListener('click', () => {
-            const fullWalletAddress = walletAddressElement.getAttribute('data-full-address'); // Get full address
-
-            if (fullWalletAddress) {
-                navigator.clipboard.writeText(fullWalletAddress)
-                .then(() => {
-                    copyMessage.style.display = 'inline';
-                    setTimeout(() => {
-                    copyMessage.style.display = 'none';
-                    }, 1000);
-                })
-                .catch(err => {
-                    console.error('Could not copy text: ', err);
-                });
-            }
-        });
-    }
+    handleCopyWalletAddress()
   
     if (usernameElement && walletAddressElement) {
         // Fetch updated profile
@@ -246,31 +164,12 @@ async function lockWallet() {
         fetchQuickLinks();
         fetchRecentServices();
         // Fetch transaction history
-        await fetchAndUpdateTransactionHistory(pageSize=5);
+        const pageSize = 5;
+        await fetchAndUpdateTransactionHistory(pageSize);
   
         // Periodic balance update
     }
-    // Logout functionality
-    const lockButton = document.getElementById('lock-wallet-btn');
-    if (lockButton) {
-        lockButton.addEventListener('click', () => {
-            const lockModal = new bootstrap.Modal(document.getElementById('exampleModal'));
-            lockModal.show();
-            const confirmButton = document.querySelector('.yes-btn');
-            const cancelButton = document.querySelector('.no-btn');
-            cancelButton.addEventListener('click', () => {
-                const modalElement = document.getElementById("exampleModal"); // Replace with your modal ID
-                modalElement.addEventListener("hidden.bs.modal", function () {
-                    document.querySelectorAll(".modal-backdrop").forEach(backdrop => backdrop.remove());
-                    document.body.classList.remove("modal-open"); // Ensure scrolling is re-enabled
-                });
-            })
-            confirmButton.addEventListener('click', () => {
-                lockModal.hide();
-                lockWallet();
-            }, { once: true });
-        });
-    }
+    handleLogout();
   });
 
 async function fetchTransactionCount() {
@@ -516,53 +415,4 @@ async function fetchQuickLinks() {
             }
         }
     });
-}
-
-// Function to load Navbar & Sidebar dynamically
-async function loadNavbarAndSidebar() {
-    let sidebarContainer = document.getElementById("sidebar-container");
-    await Promise.all([
-        fetch('navbar.html')
-            .then(response => response.text())
-            .then(html => {
-                document.getElementById('navbar-container').innerHTML = html;
-            }),
-        
-        fetch("sidebar.html")
-            .then(response => {
-                if (!response.ok) {
-                    throw new Error(`HTTP error! Status: ${response.status}`);
-                }
-                return response.text();
-            })
-            .then(data => {
-                sidebarContainer.innerHTML = data;
-        
-                // Now dynamically highlight the active menu item
-                let currentPage = window.location.pathname.split("/").pop();
-        
-                let links = {
-                    "profile.html": "dashboard-link",
-                    "transactions.html": "transactions-link",
-                    "connectedSites.html": "linked-sites-link"
-                };
-        
-                document.querySelectorAll(".nav-link").forEach(link => {
-                    link.classList.remove("active");
-                    let arrow = link.querySelector(".arrow-icon");
-                    if (arrow) arrow.style.display = "none";
-                });
-        
-                if (links[currentPage]) {
-                    let activeLink = document.getElementById(links[currentPage]);
-                    if (activeLink) {
-                        activeLink.classList.add("active");
-                        let arrow = activeLink.querySelector(".arrow-icon");
-                        if (arrow) arrow.style.display = "block";
-                    }
-                }
-            })
-        .catch(error => console.error("Error loading sidebar:", error))
-      
-    ]);
 }
